+++ src/DropdownLayout/docs/storySettings.js
@@ -1,0 +1,11 @@
+import { Category } from '../../../stories/storiesHierarchy';
+
+export const storySettings = {
+  storyName: 'DropdownLayout',
+  indexCategory: Category.PICKERS_AND_SELECTORS,
+  dataHook: 'dropdown-layout-tabs-switches-test',
+};
+
+export const testStories = {
+  tabsSwitches: 'Tabs switches',
+};