+++ src/FilePicker/FilePicker.e2e.js
@@ -6,7 +6,7 @@ import { storySettings } from './docs/storySettings';
 import { createStoryUrl } from '../../test/utils/storybook-helpers';
 
 const storyUrl = createStoryUrl({
-  kind: storySettings.kind,
+  kind: storySettings.category,
   story: storySettings.storyName,
 });