+++ src/FilePicker/docs/storySettings.js
@@ -1,9 +1,7 @@
 import { Category } from '../../../stories/storiesHierarchy';
 
 export const storySettings = {
-  kind: Category.INPUTS,
-  category: Category.INPUTS,
-
-  storyName: '3.10 + FilePicker',
+  category: Category.COMPONENTS,
+  storyName: 'FilePicker',
   dataHook: 'story-filepicker',
 };