+++ src/InputWithOptions/docs/storySettings.js
@@ -1,0 +1,11 @@
+import { Category } from '../../../stories/storiesHierarchy';
+
+export const storySettings = {
+  storyName: 'InputWithOptions',
+  category: Category.COMPONENTS,
+  dataHook: 'input-with-options-tabs-switches-test',
+};
+
+export const testStories = {
+  tabsSwitches: 'Tabs switches',
+};