+++ src/index.js
@@ -117,3 +117,4 @@ export { default as NoBorderInput } from './NoBorderInput';
 export { default as NumberInput } from './NumberInput';
 export { default as ColorInput } from './ColorInput';
 export { default as DateInput } from './DateInput';
+export { default as EditableTitle } from './EditableTitle';